Finish renaming rules to detectors, including packages, modules and so
