+++ app/build.gradle
@@ -25,7 +25,7 @@ configurations {
 }
 
 dependencies {
-    lintChecks project(path: ":rules", configuration: 'lintChecks')
+    lintChecks project(configuration: 'lintChecks', path: ':detectors')
     lintClassPath project(":dsl")
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.3.10"
