+++ detectors/.gitignore
