+++ detectors/build.gradle
@@ -1,0 +1,43 @@
+apply plugin: 'jacoco'
+apply plugin: 'kotlin'
+apply plugin: 'kotlin-kapt'
+
+sourceSets {
+    main {
+        java {
+            srcDirs = ["src/main/kotlin", "${buildDir.absolutePath}/generated/source/kapt/main/"]
+        }
+    }
+}
+
+configurations {
+    lintChecks
+}
+
+dependencies {
+    compileOnly project(":dsl")
+    compileOnly project(":annotations")
+    kapt project(":processor")
+
+    compileOnly 'com.android.tools.lint:lint:26.2.1'
+    compileOnly 'com.android.tools.lint:lint-api:26.2.1'
+    compileOnly 'com.android.tools.lint:lint-checks:26.2.1'
+
+    testCompile 'junit:junit:4.12'
+    testCompile project(":dsl")
+    testCompile 'org.assertj:assertj-core:3.3.0'
+    testCompile 'org.mockito:mockito-core:2.23.0'
+    testCompile 'com.android.tools.lint:lint:26.2.1'
+    testCompile 'com.android.tools.lint:lint-tests:26.2.1'
+    testCompile 'com.android.tools:testutils:26.2.1'
+
+    lintChecks files(jar)
+}
+
+test.finalizedBy jacocoTestReport
+
+jar {
+    manifest {
+        attributes("Lint-Registry-v2": "com.serchinastico.lin.detectors.LinIssueRegistry")
+    }
+}
