+++ detectors/src/main/kotlin/com/serchinastico/lin/detectors/JavaFileToAvoidCompilationFailure.java
@@ -1,0 +1,1 @@
+package com.serchinastico.lin.detectors;
