--- rules/src/main/kotlin/com/serchinastico/lin/rules/JavaFileToAvoidCompilationFailure.java
@@ -1,1 +1,0 @@
-package com.serchinastico.lin.rules;
