+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':dsl', ':rules', ':annotations', ':processor'
+include ':app', ':dsl', ':detectors', ':annotations', ':processor'
